src/s3_to_redshift_lambda/requirements.txt
pyarrow
psycopg2-binary
pandas
sqlalchemy